refactor: move `detectJestVersion` into `utils` folder
