+++ src/rules/no-deprecated-functions.ts
@@ -1,6 +1,10 @@
 import { AST_NODE_TYPES, TSESTree } from '@typescript-eslint/utils';
-import { JestVersion, detectJestVersion } from './detectJestVersion';
-import { createRule, getNodeName } from './utils';
+import {
+  JestVersion,
+  createRule,
+  detectJestVersion,
+  getNodeName,
+} from './utils';
 
 interface ContextSettings {
   jest?: EslintPluginJestSettings;
