+++ src/rules/utils/detectJestVersion.ts
